backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    # Use SQLite for development (easier setup)
    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL', 'sqlite:///bookish.db')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY', 'your-secret-key')